requirements.txt
numpy==1.23.5
pandas==1.5.2
beautifulsoup4==4.11.1 
pytesseract==0.3.10
spacy==3.4.4 
scikit-learn==1.2.0 
nlpaug==1.1.11 
snorkel==0.9.9 
treedlib==0.1.3 
numbskull==0.1.1
tensorflow==2.11.0
tensorboard==2.11.2 
nltk==3.8.1 
matplotlib==3.6.3 
wget==3.2
utils==1.0.1
textblob==0.17.1 
requests==2.28.2 
scrapy==2.7.1 
gensim==4.3.0 
psutil==5.9.4 
tornado==6.2
fasttext==0.9.2
pytorch-pretrained-bert==0.6.2
pytorch-nlp==0.5.0
tqdm==4.64.1 
ktrain==0.33.1
lime==0.2.0.1 
seaborn==0.11.1
shap==0.41.0 
ipython==8.9.0 
fastai==2.7.10 
torchtext==0.14.1 
transformers==4.26.1 
textacy==0.12.0 
networkx==2.8.7
html5lib==1.1 
sumy==0.11.0 
summa==1.2.0 
bert-extractive-summarizer==0.10.1 